requirements.txt
# conda create --name scout python=3.11
uv==0.7.3
google-generativeai==0.8.5
google-cloud-aiplatform==1.92.0
google-cloud-storage==2.14.0
python-dotenv==1.1.0
streamlit==1.30.0
# Cloud Run requirements
flask==2.3.3
gunicorn==21.2.0